backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
djoser==2.2.0
django-filter==23.2
django-colorfield==0.8.0
drf-extra-fields==3.4.1
drf-yasg==1.21.5
django-rest-swagger==2.2.0
gunicorn==20.1.0
python-dotenv==1.0.0
asgiref==3.6.0